Add tht_review option to clone ansible/clone.yml
Rename development option to ssh_git_checkout because that's
what the code does based on that option. The idea is that you
don't check out over SSH if ansible is doing the checkout because
ansible won't be using my private ssh key for git checkouts.
Instead with this option it downloads via HTTP.

The tht_review variable can be set to a URL you can get from
a gerrit review if you click download and then tgz next to
archive.

In this case I'm automating testing with an unmerged patch I
need for the ceph scenario to work.